requirements.txt
flask
faster-whisper
requests
gTTS

torch
soundfile
Pillow
transformers
